feat: logout & refactor: header layout
