+++ package.json
@@ -79,6 +79,7 @@
     "react-bootstrap": "^2.9.0",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
+    "react-error-boundary": "^4.0.11",
     "react-hook-form": "^7.47.0",
     "react-hot-toast": "^2.4.1",
     "react-id-swiper": "^4.0.0",
