+++ resources/app/components/guard/auth-guard.tsx
@@ -1,5 +1,7 @@
 import '@/bootstrap';
 import { useAppSelector } from '@/redux/hook';
+import { load } from '@/utilities/local-storage';
+import { storageKeys } from '@/utilities/local-storage/storage-keys';
 import React, { FC } from 'react';
 import { Navigate } from 'react-router-dom';
 type AuthGuardProps = {
@@ -9,10 +11,10 @@ type AuthGuardProps = {
 export const AuthGuard: FC<AuthGuardProps> = ({ children }) => {
   const { accessToken } = useAppSelector(state => state.auth);
 
+  const isStayIn = load(storageKeys.STAY_IN);
   // render
-  if (!accessToken) {
+  if (!accessToken && !isStayIn) {
     return <Navigate to="/login" />;
-  } else {
-    return <>{children}</>;
   }
+  return <>{children}</>;
 };
