+++ resources/app/pages/auth/account-setting/components/form-reset.tsx
