+++ resources/app/pages/auth/privacy-security/index.tsx
@@ -1,0 +1,14 @@
+import { Container } from 'react-bootstrap';
+
+export const PrivacySecurityPage = () => {
+  return (
+    <>
+      <div id="content-page" className="content-page">
+        <Container>
+          <h1>Privacy - Security </h1>
+          <h2>We are updating, Coming soon... </h2>
+        </Container>
+      </div>
+    </>
+  );
+};
