+++ resources/app/utilities/local-storage/storage-keys.ts
@@ -3,6 +3,7 @@ export const storageKeys = {
   USER_ID: 'user_id',
   ACCESS_TOKEN: 'access_token',
   REFRESH_TOKEN: 'refresh_token',
+  STAY_IN: 'stay_in',
 } as const;
 export type StorageKeyType = keyof typeof storageKeys;
